Leverage PR #5251 to bring back trampoline_self_life_support.h from smart_holder branch almost as-is.
This ensures IWYU correctness in client code.

The only change in trampoline_self_life_support.h is to replace `#include "detail/type_caster_base.h"` with "#include detail/value_and_holder.h".
